Splitting models.py in to seperate files, each containing one class
