+++ backend/custom_models/item_text.py
@@ -1,0 +1,14 @@
+from django.db import models
+from .item import Item
+from .language import Language
+MAX_CHAR_LEN = 100
+
+
+class ItemText(models.Model):
+    item_text_id = models.AutoField(primary_key=True)
+    item_id = models.ForeignKey(Item, on_delete=models.CASCADE)
+    text_detail = models.CharField(max_length=MAX_CHAR_LEN)
+    language = models.ForeignKey(Language, on_delete=models.CASCADE)
+    date_created = models.DateTimeField()
+    date_from = models.DateTimeField()
+    date_to = models.DateTimeField(null=True, blank=True)
